+++ packages/gha-core/src/error/index.ts
@@ -1,0 +1,2 @@
+export * from './class';
+export * from './handler';
