+++ packages/gha-core/src/error/message.ts
@@ -1,0 +1,1 @@
+export {};
