--- packages/gha-libs/src/error/index.ts
@@ -1,1 +1,0 @@
-export * from './class';
