+++ packages/gha-libs/src/index.ts
@@ -1,3 +1,2 @@
 export * from './configs';
 export * from './decorator';
-export * from './error';
